move patterninputhist to its own file
